card played when dropped in middle. card returns to hand if not dropped in middle
