+++ client/lib/Deck.js
@@ -1,0 +1,9 @@
+export default class Deck {
+  constructor() {
+    this.cardsDealt = [];
+  }
+
+  dealCard(card) {
+    this.cardsDealt.push(card);
+  }
+}
